requirements.txt
python-dotenv==1.0.1
langchain==0.3.18
langchain-community==0.3.17
# langchain-core==0.3.35
tiktoken==0.9.0
sentence-transformers==2.3.1
tokenizers==0.21.0
# faiss-cpu===1.10.0
unstructured==0.16.23
# numpy===1.26.4
nltk==3.9.1
-e .
fastapi==0.115.8
uvicorn==0.34.0
jinja2==3.1.5
python-multipart==0.0.20

## Llama 2 with Pinecone requirements ##
# langchain - Already added
# sentence_transformers - Already added
# unstructured - Already added
# tokenizers - Already added
pinecone-client==3.0.2
bitsandbytes==0.41.1
accelerate==0.26.1
transformers==4.49.0
datasets==2.17.1
loralib==0.1.2
sentencepiece==0.1.99
pypdf==4.0.1
chromadb==0.6.3
torch==2.2.0
